[funct] add logout function
-add function logout
-add api endpoint
-add respective controller method

commited by yahia Alioua

+++ AngularAuthApi/Authentication/Abstractions/ITokenGenerator.cs
@@ -4,6 +4,6 @@ namespace AngularAuthApi.Authentication
 {
     public interface ITokenGenerator
     {
-        string GenerateToken(string secretKey, string issuer, string audience, int expires, List<Claim> claims = null);
+        string GenerateToken(string secretKey, string issuer, string audience, double expires, List<Claim> claims = null);
     }
 }
