+++ AngularAuthApi/Authentication/JwtConfigOptions.cs
@@ -3,7 +3,7 @@
     public class JwtConfigOptions
     {
         public string AccessTokenSecretKey { get; init; }
-        public int AccessTokenExpirationMinutes { get; set; }
+        public double AccessTokenExpirationMinutes { get; set; }
         public string ?Issuer { get; set; }
         public string ?Audience { get; set; }
