+++ AngularAuthApi/Authentication/OptionsSetup/RefreshJwtConfigOptions.cs
@@ -3,7 +3,7 @@
     public class RefreshJwtConfigOptions
     {
         public string RefreshTokenSecretKey { get; init; }
-        public int RefreshTokenExpirationHours { get; set; }
+        public double RefreshTokenExpirationHours { get; set; }
         public string Issuer { get; set; }
         public string Audience { get; set; }
     }
